pyproject.toml
[tool.poetry]
name = "rbf"
version = "0.1.0"
description = "Reddit Bot Framework"
authors = ["jmhayes3 <jmhayes3@jmhayes3.github.com>"]
repository = "https://github.com/jmhayes3/rbf"
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.3"
pyzmq = "^25.1.0"
praw = "^7.7.0"
sqlalchemy = "^2.0.15"

[tool.poetry.group.dev.dependencies]
pylint = "^2.17.4"

[tool.poetry.scripts]
rbf-server = "rbf.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
